# openapi/mobile-v1.yaml
openapi: 3.0.3
info:
  title: ChatSwitch API
  version: "1.0.0"
  description: |
    ## Мобильный REST API (`/api/v1/...`)
    Bearer-токен Laravel Sanctum. Роли `administrator`, `manager`, `employee`.

    **Неактивный пользователь** (`is_active: false`): логин — HTTP 403; при уже выданном токене — 403.

    **Лимиты:** `throttle:api`; неудачные логины — до 5 попыток по email+IP (429).

    ---

    ## Веб-эндпоинты (сессия, CSRF)
    Маршруты для оператора браузера: AI-ассистент, воронки продаж, задачи по отделам.
    Требуют cookie-сессии (`auth`) и заголовка `X-CSRF-TOKEN` / `X-Requested-With: XMLHttpRequest`.
    Роль — через middleware `role:administrator` (там, где указано).
servers:
  - url: /
    description: Текущий хост приложения
...
paths:
...
  /api/v1/chats:
    get:
      tags: [Chats]
      summary: Список чатов (не в архиве)
      operationId: chatsIndex
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: per_page
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - in: query
          name: search
          schema: { type: string }
          description: Поиск по имени чата / контакту (как на вебе)
      responses:
        "200":
          description: Пагинированный список
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedChats"
        "401":
          description: Не авторизован
        "403":
          description: Нет доступа
...
  /api/v1/chats/{chat}/messages:
    get:
      tags: [Chats]
      summary: История сообщений (курсор)
      operationId: chatMessages
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ChatId"
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - in: query
          name: before_timestamp
          schema: { type: string }
          description: Загрузка страницы старше этого времени
        - in: query
          name: before_id
          schema: { type: integer }
          description: Уточнение при равном before_timestamp
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesCursorResponse"
        "403":
          description: Нет view на чат
    post:
      tags: [Chats]
      summary: Отправить исходящее текстовое сообщение
      operationId: chatSendMessage
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ChatId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message:
                    $ref: "#/components/schemas/Message"
        "403":
          description: Нет sendMessage / view
        "422":
          description: Валидация
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum personal access token
      description: 'Заголовок Authorization: Bearer {token}'
...
  parameters:
    ChatId:
      name: chat
      in: path
      required: true
      description: ID чата
      schema:
        type: integer
...
  schemas:
...
    Chat:
      type: object
      description: Поля из ChatResource (часть опциональна, если связь не загружена)
      properties:
        id: { type: integer }
        whatsapp_chat_id: { type: string }
        whatsapp_session_id: { type: integer }
        contact_id: { type: integer, nullable: true }
        chat_name: { type: string, nullable: true }
        is_group: { type: boolean }
        last_message_text: { type: string, nullable: true }
        last_message_at: { type: string, format: date-time, nullable: true }
        last_message_direction: { type: string, nullable: true }
        unread_count: { type: integer }
        is_archived: { type: boolean }
        is_pinned: { type: boolean }
        is_muted: { type: boolean }
        muted_until: { type: string, format: date-time, nullable: true }
        is_favorite: { type: boolean }
        pinned_message_id: { type: integer, nullable: true }
    PaginatedChats:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Chat"
        links:
          type: object
          additionalProperties: true
        meta:
          type: object
          additionalProperties: true
    Message:
      type: object
      description: Поля из MessageResource
      properties:
        id: { type: integer }
        chat_id: { type: integer }
        direction: { type: string, enum: [inbound, outbound, system] }
        type: { type: string }
        body: { type: string, nullable: true }
        metadata: { type: object, additionalProperties: true, nullable: true }
        sender_phone: { type: string, nullable: true }
        sender_name: { type: string, nullable: true }
        sent_by_user_id: { type: integer, nullable: true }
        is_forwarded: { type: boolean }
        quoted_message_id: { type: string, nullable: true }
        ack: { type: string }
        message_timestamp: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time, nullable: true }
    MessagesCursorResponse:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
    SendMessageRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          maxLength: 4096
        display_message:
          type: string
          maxLength: 4096
          nullable: true
        quoted_message_id:
          type: string
          maxLength: 255
          nullable: true
        mentions:
          type: array
          maxItems: 20
          items: { type: string }
        mentions_meta:
          type: array
          maxItems: 20
          items:
            type: object
            required: [id, number, label]
            properties:
              id: { type: string }
              number: { type: string }
              label: { type: string }
...